automatically detect common base path

+++ src/lib/fs.js
@@ -1,10 +1,19 @@
 import fs from 'fs-extra'
+import path from 'path'
+
+const cwd = process.cwd()
 
 export const copy = (source, target, notify) => {
   notify('copy', [source, target])
 
   try {
-    fs.copySync(source, target)
+    const sourceStat = fs.statSync(source)
+
+    if (sourceStat.isFile()) {
+      fs.copySync(source, target)
+    } else if (sourceStat.isDirectory()) {
+      fs.ensureDirSync(target)
+    }
     return true
   } catch (e) {
     notify('error', e)
@@ -30,4 +39,13 @@ export const deleteExtra = (fileordir, options, notify) => {
     notify('no-delete', fileordir)
     return true
   }
+}
+
+const getBase = (file, base) => base.filter((base) => file.indexOf(base) !== -1)
+    .reduce((hit, base) => hit.length > base.length ? hit : base, '')
+
+export const resolveTarget = (file, target, options) => {
+  const from = path.join(cwd, getBase(file, options.base))
+
+  return path.join(target, path.relative(from, file))
 }
